backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
google-generativeai>=0.7.0
python-docx>=1.1.0
PyPDF2>=3.0.0
openpyxl>=3.1.2
docx2pdf>=0.1.8